Groups list page phase one complete.
