+++ app/src/main/res/menu/activity_main_button.xml
@@ -1,12 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_one"
-        android:icon="@drawable/ic_next"
-        android:orderInCategory="1"
-        android:title="Choice 1"
-        app:showAsAction="always" />
-
 </menu>
